pyproject.toml
[tool.poetry]
name = "evolution_game"
version = "0.1.0"
description = ""
authors = ["Jagath Jai Kumar <jagath.jaikumar@gmail.com>"]
readme = "README.md"
packages = [
    { include = "*", from = "evolution" }
]

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.4"
dj-database-url = "^2.3.0"
django-environ = "^0.11.2"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-extensions = "^3.2.3"
python-jose = "^3.3.0"
drf-spectacular = "^0.28.0"
django-debug-toolbar = "^5.0.1"
requests = "^2.32.3"

[tool.poetry.group.deploy.dependencies]
gunicorn = "^23.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.1"

[tool.poetry.group.test.dependencies]
pytest-django = "^4.9.0"

[tool.poetry.group.client.dependencies]
evolution-client = { path = "src/evolution-client", develop = true }

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120